+++ classes/db.py
@@ -1,16 +1,13 @@
 # Copyright 2022-TODAY Francesco Apruzzese <cescoap@gmail.com>
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl).
 
-from random import randint
 from json import dumps, loads
 import sqlite3
 from os.path import join
 
 from consts.consts import \
     VERSION, BUILD, \
-    ELEMENTS, RESOURCES, \
-    ROWS_LIMIT, ROWS_NUMBER, \
-    COLS_LIMIT, COLS_NUMBER
+    ELEMENTS, RESOURCES
 
 
 class Db:
@@ -107,6 +104,8 @@ class Db:
             "production TEXT, "
             "production_on_build TEXT, "
             "building_constraints TEXT, "
+            "consumption TEXT, "
+            "recovery_on_demolish TEXT, "
             "block INTEGER, "
             "symbol TEXT NOT NULL UNIQUE, "
             "emoji TEXT"
@@ -174,12 +173,18 @@ class Db:
                     element.get('production_on_build'))
                 building_constraints = self._from_dict_to_database(
                     element.get('building_constraints'))
+                consumption = self._from_dict_to_database(
+                    element.get('consumption'))
+                recovery_on_demolish = self._from_dict_to_database(
+                    element.get('recovery_on_demolish'))
                 self.cursor.execute(
                     "INSERT OR IGNORE INTO element ("
                     "code, category, name, "
                     "cost, "
                     "production, production_on_build, "
                     "building_constraints, "
+                    "consumption, "
+                    "recovery_on_demolish, "
                     "symbol, emoji, block"
                     ") VALUES ("
                     f"'{element_type}-{element_name}', "
@@ -189,6 +194,8 @@ class Db:
                     f"'{production}', "
                     f"'{production_on_build}', "
                     f"'{building_constraints}', "
+                    f"'{consumption}', "
+                    f"'{recovery_on_demolish}', "
                     f"'{element['symbol']}', "
                     f"'{element.get('emoji', '')}', "
                     f"{self._convert_in_boolean(element.get('block', False))}"