+++ classes/element.py
@@ -15,6 +15,8 @@ class Element:
     production = None
     production_on_build = None
     building_constraints = None
+    consumption = None
+    recovery_on_demolish = None
     _app = None
     _db = None
 
@@ -38,6 +40,8 @@ class Element:
         self.production = data['production']
         self.production_on_build = data['production_on_build']
         self.building_constraints = data['building_constraints']
+        self.consumption = data['consumption']
+        self.recovery_on_demolish = data['recovery_on_demolish']
         return self
 
     def _get_raw_data(self):
@@ -56,7 +60,10 @@ class Element:
                 structured_data[data] = bool(raw_data[data])
             elif data in ('cost',
                           'production', 'production_on_build',
-                          'building_constraints'):
+                          'building_constraints',
+                          'consumption',
+                          'recovery_on_demolish',
+                          ):
                 structured_data[data] = self._db._from_database_to_dict(
                     raw_data[data])
             else: